feat: Refactor and organize code structure using their class name as file name, isolate LogicExtractor.cpp/hpp
